Fix some bugs related to keyboard input handling, especially modifier keys

+++ kernel/src/devices/PS2/Keypoints.cpp
@@ -375,7 +375,7 @@ namespace {
 }
 
 namespace Devices::PS2 {
-    static constexpr uint8_t KEY_PRESSED = Devices::KeyboardDispatcher::FLAG_KEY_PRESSED;
+    static constexpr uint16_t KEY_PRESSED = Devices::KeyboardDispatcher::FLAG_KEY_PRESSED;
 
     EventResponse KeyboardScanCodeSet1Handler(uint8_t byte, BasicKeyPacket* buffer) {
         static enum class state_t {
@@ -404,7 +404,7 @@ namespace Devices::PS2 {
                 BasicKeyPacket packet = {
                     .scancode = byte,
                     .keypoint = scan_code_set_1_keypoints[byte],
-                    .flags = static_cast<uint8_t>(byte & 0x80 ? 0 : KEY_PRESSED)
+                    .flags = static_cast<uint16_t>(byte & 0x80 ? 0 : KEY_PRESSED)
                 };
                 *buffer = packet;
                 return EventResponse::PACKET_CREATED;
@@ -426,7 +426,7 @@ namespace Devices::PS2 {
                 BasicKeyPacket packet = {
                     .scancode = byte,
                     .keypoint = ext_1_scan_code_set_1_keypoints[byte],
-                    .flags = static_cast<uint8_t>(byte & 0x80 ? 0 : KEY_PRESSED)
+                    .flags = static_cast<uint16_t>(byte & 0x80 ? 0 : KEY_PRESSED)
                 };
                 *buffer = packet;
                 return EventResponse::PACKET_CREATED;
@@ -439,7 +439,7 @@ namespace Devices::PS2 {
                     BasicKeyPacket packet = {
                         .scancode = byte,
                         .keypoint = ext_2_scan_code_set_1_keypoint,
-                        .flags = static_cast<uint8_t>(KEY_PRESSED)
+                        .flags = KEY_PRESSED
                     };
                     *buffer = packet;
                     return EventResponse::PACKET_CREATED;
@@ -463,7 +463,7 @@ namespace Devices::PS2 {
                     BasicKeyPacket packet = {
                         .scancode = byte,
                         .keypoint = ext_2_scan_code_set_1_keypoint,
-                        .flags = static_cast<uint8_t>(0)
+                        .flags = 0
                     };
                     *buffer = packet;
                     return EventResponse::PACKET_CREATED;
@@ -487,7 +487,7 @@ namespace Devices::PS2 {
                     BasicKeyPacket packet = {
                         .scancode = byte,
                         .keypoint = ext_3_scan_code_set_1_keypoint,
-                        .flags = static_cast<uint8_t>(KEY_PRESSED)
+                        .flags = KEY_PRESSED
                     };
                     *buffer = packet;
                     return EventResponse::PACKET_CREATED;
@@ -541,7 +541,7 @@ namespace Devices::PS2 {
                 BasicKeyPacket packet = {
                     .scancode = byte,
                     .keypoint = scan_code_set_2_keypoints[byte % 0x84],
-                    .flags = static_cast<uint8_t>(onRelease ? 0 : KEY_PRESSED)
+                    .flags = static_cast<uint16_t>(onRelease ? 0 : KEY_PRESSED)
                 };
                 onRelease = 0;
                 *buffer = packet;
@@ -569,7 +569,7 @@ namespace Devices::PS2 {
                 BasicKeyPacket packet = {
                     .scancode = byte,
                     .keypoint = ext_1_scan_code_set_2_keypoints[byte],
-                    .flags = static_cast<uint8_t>(onRelease ? 0 : KEY_PRESSED)
+                    .flags = static_cast<uint16_t>(onRelease ? 0 : KEY_PRESSED)
                 };
                 onRelease = 0;
                 *buffer = packet;
@@ -583,7 +583,7 @@ namespace Devices::PS2 {
                     BasicKeyPacket packet = {
                         .scancode = byte,
                         .keypoint = ext_2_scan_code_set_2_keypoint,
-                        .flags = static_cast<uint8_t>(KEY_PRESSED)
+                        .flags = KEY_PRESSED
                     };
                     *buffer = packet;
                     return EventResponse::PACKET_CREATED;
@@ -609,7 +609,7 @@ namespace Devices::PS2 {
                     BasicKeyPacket packet = {
                         .scancode = byte,
                         .keypoint = ext_2_scan_code_set_2_keypoint,
-                        .flags = static_cast<uint8_t>(0)
+                        .flags = 0
                     };
                     *buffer = packet;
                     return EventResponse::PACKET_CREATED;
@@ -634,7 +634,7 @@ namespace Devices::PS2 {
                     BasicKeyPacket packet = {
                         .scancode = byte,
                         .keypoint = ext_3_scan_code_set_2_keypoint,
-                        .flags = static_cast<uint8_t>(KEY_PRESSED)
+                        .flags = KEY_PRESSED
                     };
                     *buffer = packet;
                     return EventResponse::PACKET_CREATED;
@@ -669,7 +669,7 @@ namespace Devices::PS2 {
         BasicKeyPacket packet = {
             .scancode = byte,
             .keypoint = scan_code_set_3_keypoints[byte],
-            .flags = static_cast<uint8_t>(state == state_t::RELEASE ? 0 : KEY_PRESSED)
+            .flags = static_cast<uint16_t>(state == state_t::RELEASE ? 0 : KEY_PRESSED)
         };
         state = state_t::DEFAULT;
         *buffer = packet;
